requirements.txt
black~=22.3.0
click
coloredlogs
ffmpeg-python
mypy-extensions~=0.4.3
pre-commit~=2.20.0
pyTelegramBotAPI
python-dotenv~=0.21.0
rich
sox
wget
whisper-cpp-python
